Modularizing heartbeat function
